fix(filters): reject an incomplete address in the bar
isValidAddress covered EVM, bitcoin, bitcoin cash and solana, but not the
substrate addresses the backend accepts, so a polkadot or kusama address was
refused by history's address filter.

Add isValidSs58Address, built on the base58 decoder already here, and check the
decoded length and the network prefix. The two-byte blake2b checksum is left to
the backend, as the sibling validators leave theirs.

The shared address field then validates with it. Without that a table could opt
out, as the address book did, and a half-typed address was both offered as a
suggestion and applied on closing the editor, which the backend answered with a
400.
