+++ frontend/app/src/modules/core/table/filters/shared/shared-fields.spec.ts
@@ -54,13 +54,26 @@ describe('decorateSharedField', () => {
     expect(txHash.resolveLabel?.('0xabc')).toBe('short:0xabc');
   });
 
-  // What a value means to the backend stays with the table's matcher: history's transaction filter
+  // What a hash means to the backend stays with the table's matcher: history's transaction filter
   // accepts a signature as well as a hash, which no shape check here would know about.
-  it('should leave validation to the field it decorates', () => {
+  it('should leave transaction hash validation to the field it decorates', () => {
     const validate = (value: string): boolean => value === 'only-this';
-    const decorated = decorateSharedField(field({ validate }), SharedFieldKinds.ADDRESS, resolvers);
+    const decorated = decorateSharedField(field({ validate }), SharedFieldKinds.TX_HASH, resolvers);
 
     expect(decorated.validate).toBe(validate);
+  });
+
+  // An address is different: it is sent as an address, and the backend answers anything it cannot
+  // place with a 400. A table opting out of that (the address book matcher validated nothing) meant
+  // a half-typed address was both offered as a suggestion and applied on closing the editor.
+  it('should validate an address whatever the field says', () => {
+    const decorated = decorateSharedField(field({ validate: () => true }), SharedFieldKinds.ADDRESS, resolvers);
+
+    expect(decorated.validate?.('0x')).toBe(false);
+    expect(decorated.validate?.('0x742d35Cc6634C0532925a3b844Bc9e7595f0bEbB')).toBe(true);
+    // Every ecosystem the backend accepts, not just EVM.
+    expect(decorated.validate?.('13UVJyLnbVp9RBZYFwFGyDvVd1y27Tt8tkntv6Q7JVPhFsTB')).toBe(true);
+    expect(decorated.validate?.('1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa')).toBe(true);
   });
 
   it('should keep the key, operators and binding the table declared', () => {
